+++ Anonymous_Survey_Ardalis/src/Anonymous_Survey_Ardalis.Web/Subjects/Delete/Delete.cs
@@ -1,11 +1,13 @@
+using Anonymous_Survey_Ardalis.Core.Interfaces;
 using Anonymous_Survey_Ardalis.UseCases.Subjects.Commands.Delete;
+using Anonymous_Survey_Ardalis.Web.Security;
 using Ardalis.Result;
 using FastEndpoints;
 using MediatR;
 
 namespace Anonymous_Survey_Ardalis.Web.Subjects;
 
-public class Delete(IMediator _mediator)
+public class Delete(IMediator _mediator, IAdminPermissionService _adminPermissionService, ICurrentUserService _currentUserService)
   : Endpoint<DeleteSubjectRequest>
 {
   public override void Configure()
@@ -17,6 +19,13 @@ public class Delete(IMediator _mediator)
     DeleteSubjectRequest request,
     CancellationToken cancellationToken)
   {
+    var adminId = _currentUserService.GetCurrentAdminId();
+    if (!await _adminPermissionService.CanDeleteSubject(adminId, request.Subjectid))
+    {
+      await SendForbiddenAsync();
+      return;
+    }
+
     var command = new DeleteSubjectCommand(request.Subjectid);
 
     var result = await _mediator.Send(command, cancellationToken);
